GEODE-4080: PR feedback changes
* Get rid of the codec package
* Inline the serialization type enum
* Fix a doc comment
