--- geode-client-protocol/src/main/java/org/apache/geode/internal/protocol/protobuf/v1/serializer/ProtocolSerializer.java
@@ -1,34 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more contributor license
- * agreements. See the NOTICE file distributed with this work for additional information regarding
- * copyright ownership. The ASF licenses this file to You under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance with the License. You may obtain a
- * copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- */
-package org.apache.geode.internal.protocol.protobuf.v1.serializer;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-
-import org.apache.geode.annotations.Experimental;
-import org.apache.geode.internal.protocol.exception.InvalidProtocolMessageException;
-
-/**
- * This interface is used to translate between binary data and protocol specific messages.
- *
- * @param <T> The message type of the protocol.
- */
-@Experimental
-public interface ProtocolSerializer<T> {
-  T deserialize(InputStream inputStream) throws InvalidProtocolMessageException;
-
-  void serialize(T inputMessage, OutputStream outputStream) throws IOException;
-}
